requirements.txt
Flask==3.0.0
google.cloud==0.34.0
google.cloud.datastore==2.19.0
google.appengine.ext==*
protobuf==3.20.*